+++ inst/tinytest/cpp/test-vectors.cpp
@@ -2,18 +2,12 @@
 //
 // This file is part of RcppBlaze.
 //
+// This file is part of RcppBlaze.
+//
 // RcppBlaze is free software: you can redistribute it and/or modify it
-// under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 2 of the License, or
-// (at your option) any later version.
-//
-// RcppBlaze is distributed in the hope that it will be useful, but
-// WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with RcppBlaze If not, see <http://www.gnu.org/licenses/>.
+// under the terms of the 3-Clause BSD License. You should have received
+// a copy of 3-Clause BSD License along with RcppBlaze.
+// If not, see https://opensource.org/license/BSD-3-Clause.
 
 // [[Rcpp::depends(RcppBlaze)]]
 #include <RcppBlaze.h>