+++ NEWS.md
@@ -3,6 +3,7 @@
 * Update Blaze-lib to 3.8.2 since Rcpp supports C++14 / C++17 since 4.2.0.
 * Support `dgTMatrix` thanks to robinsonjj's contribution.
 * Enable C++11 Threads when compiling `fastLm`.
+* Change to use BSD License which is aligned with `blize-lib`.
 
 # RcppBlaze v0.2.2